Refact: single modal component for saving requests
